requirements.txt
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.10.4
altair==4.2.2
ansicon==1.89.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
banal==1.0.6
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
bokeh==2.4.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
croniter==1.3.14
cycler==0.11.0
dataset==1.6.0
datasets==2.11.0
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.2.0
fastapi==0.88.0
fastjsonschema==2.16.3
ffmpy==0.3.0
filelock==3.9.0
Flask==2.3.2
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
gradio==3.28.1
gradio_client==0.1.4
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.14.1
idna==3.4
inquirer==3.1.3
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
kiwisolver==1.4.4
lightning==2.0.2
lightning-cloud==0.5.34
lightning-utilities==0.8.0
linkify-it-py==2.0.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
nbclassic==0.5.5
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.1
openai==0.27.6
ordered-set==4.1.0
orjson==3.8.11
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.3.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pyct==0.5.0
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
pyviz-comms==2.2.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
readchar==4.0.5
regex==2023.3.23
requests==2.28.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.4
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
Send2Trash==1.8.0
sentry-sdk==1.21.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.47
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
toolz==0.12.0
torchmetrics==0.11.4
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typing_extensions==4.4.0
tzdata==2023.3
uc-micro-py==1.0.2
uri-template==1.2.0
urllib3==1.26.13
uvicorn==0.21.1
wandb==0.15.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.3.4
widgetsnbextension==4.0.7
xxhash==3.2.0
y-py==0.5.9
yarl==1.9.1
ypy-websocket==0.8.2
